+++ courier/template/template.go
@@ -12,19 +12,7 @@ import (
 	"github.com/ory/x/httpx"
 )
 
-type (
-	Config interface {
-		CourierTemplatesRoot() string
-		CourierTemplatesVerificationInvalid() *config.CourierEmailTemplate
-		CourierTemplatesVerificationValid() *config.CourierEmailTemplate
-		CourierTemplatesRecoveryInvalid() *config.CourierEmailTemplate
-		CourierTemplatesRecoveryValid() *config.CourierEmailTemplate
-		CourierTemplatesLoginValid() *config.CourierEmailTemplate
-		CourierTemplatesRegistrationValid() *config.CourierEmailTemplate
-	}
-
-	Dependencies interface {
-		CourierConfig() config.CourierConfigs
-		HTTPClient(ctx context.Context, opts ...httpx.ResilientOptions) *retryablehttp.Client
-	}
-)
+type Dependencies interface {
+	CourierConfig() config.CourierConfigs
+	HTTPClient(ctx context.Context, opts ...httpx.ResilientOptions) *retryablehttp.Client
+}