+++ identity/address.go
@@ -5,5 +5,4 @@ package identity
 
 const (
 	AddressTypeEmail = "email"
-	AddressTypePhone = "phone"
 )