+++ identity/credentials_code.go
@@ -11,7 +11,6 @@ type CodeAddressType string
 
 const (
 	CodeAddressTypeEmail CodeAddressType = AddressTypeEmail
-	CodeAddressTypePhone CodeAddressType = AddressTypePhone
 )
 
 // CredentialsCode represents a one time login/registration code