+++ persistence/sql/migrations/sql/20231130094628000000_courier_message_channel.down.sql
@@ -1,0 +1,2 @@
+ALTER TABLE
+  courier_messages DROP column channel;