+++ persistence/sql/migrations/sql/20231130094628000000_courier_message_channel.up.sql
@@ -1,0 +1,4 @@
+ALTER TABLE
+  courier_messages
+ADD
+  column channel VARCHAR(32) NULL;